Fix immediate auth failure in case of missing credential store
If there is a credential store setup for the connection,
keyboard-interactive auth method will be used to inject the password
to the target server (opaquely to the user behind the client).

If that fails (because there is no cred store set up), it will change
to password auth method opaquely. This was the point where we had
authentication failures before.

At this point Guacamole can ask for the password with a prompt and then
use not keyboard-interactive but password auth mode of SSH server to
authorize with the interactively gotten password from the user.