GUACAMOLE-377: Remove use of END_FRAME operation from guac_display.
The END_FRAME operation was previously used to notify workers that the
frame has ended, but since the receiving worker needs to check and push
that operation back onto the queue if other workers are still busy, it's
essentially unnecessary _and_ results in several workers spinning as
they pass END_FRAME around until all others are done.

It's sufficient to simply check whether the operation queue is empty and
no other workers are active.