GUACAMOLE-2153: Fix GUAC_*_UTF16 on big-endian machine
Since UTF-16 works with 16-bit values it has a byte order and that byte
order is not well defined. There is a convention to use U+FEFF as first
character which allows determining the endianess but it's commonly not
present and doesn't seem to fit into the current code. The most common
thing is to assume little-endian which is what I did here and is already
implied by the test suite. If supporting both big- and little-endian
UTF-16 is wanted one could create _UTF16LE and _UTF16BE variations of
the functions.